requirements.txt
Django==4.1
djangorestframework==3.13.1
pytest-django==4.5.2
Faker==13.15.1
black==22.6.0
pylint==2.14.5
pylint-django==2.5.3
drf-yasg==1.21.3
factory-boy==3.2.1
pycountry==22.3.5